+++ Templates/Tests/Context/AutoMockable.swift
@@ -18,6 +18,10 @@ protocol BasicProtocol: AutoMockable {
     func save(configuration: String)
 }
 
+protocol ImplicitlyUnwrappedOptionalReturnValueProtocol: AutoMockable {
+  func implicitReturn() -> String!
+}
+
 protocol InitializationProtocol: AutoMockable {
     init(intParameter: Int, stringParameter: String, optionalParameter: String?)
     func start()