+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Sourcery CHANGELOG
 
 ---
-## Master
+## 0.16.0
 
 - Replaces environment variables inside .yml configurations (like ${PROJECT_NAME}), if a value is set.
 - Fixes warning in generated AutoMockable methods that have implicit optional return values